backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
supabase==2.4.2
google-generativeai
langchain
langchain-google-genai
langchain-community
pypdf==4.1.0
python-docx==1.1.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0